docs: improve go docs of writer implementation
This commit improves the go documentation of the
`EncWriter` and `DecWriter` structs and the
`Write*` and `ReadFrom` implementations.

In general, the docs are now more expressive and explicit
about the invariants and error conditions.

Further, this commit makes the `ReadFrom` implementations
compliant to its specification. Before, `ReadFrom` closed
the `EncWriter` / `DecWriter`.