fix incorrect ReadFrom benchmarks
This commit fixes a bug in the `ReadFrom` benchmark implementation
for encryption. The `EncWriter` is not closes such that the last
fragment (resp. the only fragment) gets only copied to the in-memory
buffer but not encrypted nor written to the underlying io.Writer.

This leads to wrong benchmark results - i.e. when the data size
is smaller than the fragment size since we only measure the mem-copy.