add AES-GCM support on ppc64le and bump go version
This commit adds pcc64le to the list of architectures that
support AES-GCM - but only for Go 1.14 and newer.
Since Go 1.14 there is a asm implementation of AES-GCM.

Fixes #52

Further, this commit updates the required dependencies and
bumps the required Go version (from 1.12 to 1.13). We bump the
required Go version since Go only actively supports the latest
two releases. Since Go 1.14 is now available we drop Go 1.12.